# src/integration-test/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: api-web-message-collector
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "2.0"
servers:
  - url: http://localhost:57161
    description: Generated server url
tags:
  - name: messages
    description: Messages
paths:
  /messages:
    get:
      tags:
        - messages
      summary: Get a list of messages related to a specific familyId
      description: Returns a list of messages found for the specified familyId
      operationId: getMessages
      parameters:
        - name: familyid
          in: query
          description: FamilyId to fetch messages for
          required: true
          schema:
            type: string
          example: 123
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Problem'
                  - $ref: '#/components/schemas/ConstraintViolationProblem'
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageDTO'
    delete:
      tags:
        - messages
      summary: Delete a list of messages
      description: Deletes a list of messages with the ids provided
      operationId: deleteMessages
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int32
        required: true
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "204":
          description: No Content
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Problem'
                  - $ref: '#/components/schemas/ConstraintViolationProblem'
  /messages/attachments/{attachmentId}:
    get:
      tags:
        - messages
      summary: Get a messageAttachment
      description: Returns a messageAttachment for the specified messageId
      operationId: getAttachment
      parameters:
        - name: attachmentId
          in: path
          description: MessageId to fetch attachment for
          required: true
          schema:
            type: integer
            format: int32
          example: 123
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Problem'
                  - $ref: '#/components/schemas/ConstraintViolationProblem'
        "200":
          description: Successful Operation
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags:
        - messages
      summary: Delete a messageAttachment
      description: Deletes a messageAttachment with the specified id
      operationId: deleteAttachment
      parameters:
        - name: attachmentId
          in: path
          description: Id of the attachment to delete
          required: true
          schema:
            type: integer
            format: int32
          example: 123
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "204":
          description: No Content
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Problem'
                  - $ref: '#/components/schemas/ConstraintViolationProblem'
  /api-docs:
    get:
      tags:
        - API
      summary: OpenAPI
      operationId: getApiDocs
      responses:
        "200":
          description: OK
          content:
            application/yaml:
              schema:
                type: string
      x-auth-type: None
      x-throttling-tier: Unlimited
      x-wso2-mutual-ssl: Optional
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: '#/components/schemas/StatusType'
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: '#/components/schemas/ThrowableProblem'
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: '#/components/schemas/StatusType'
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: '#/components/schemas/ThrowableProblem'
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: '#/components/schemas/StatusType'
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    MessageAttachment:
      type: object
      properties:
        attachmentId:
          type: integer
          description: The Id for the attachment
          format: int32
          example: 1
        fileName:
          type: string
          description: The name of the file
          example: file.txt
        file:
          type: string
          description: The file as a base64 encoded string
          example: dGhpcyBpcyBhIGJhc2U2NCBlbmNvZGVkIHN0cmluZw==
      description: List of attachments for the message
      example: "attachment1, attachment2"
    MessageDTO:
      type: object
      properties:
        id:
          type: integer
          description: The webMessageCollector Id for the message
          format: int32
          example: 1
        direction:
          type: string
          description: If the message is inbound or outbound from the perspective
            of case-data/e-service.
          example: INBOUND
          enum:
            - INBOUND
            - OUTBOUND
        familyId:
          type: string
          description: What E-service the message was found in
          example: "501"
        externalCaseId:
          type: string
          description: 'The external caseID '
          example: caa230c6-abb4-4592-ad9a-34e263c2787b
        message:
          type: string
          description: 'The message '
          example: Hello World
        messageId:
          type: string
          description: The unique messageId from openE for the message
          example: "12"
        sent:
          type: string
          description: 'Time and date the message was sent '
          example: 2023-02-23 17:26:23
        username:
          type: string
          description: Username for the poster
          example: te01st
        firstName:
          type: string
          description: 'Firstname of the poster '
          example: Test
        lastName:
          type: string
          description: 'If the message was utbound from a manager or inbound from
            a customer '
          example: Testorsson
        email:
          type: string
          description: Email for the poster
          example: test@sundsvall.se
        userId:
          type: string
          description: 'If the message was utbound from a manager or inbound from
            a customer '
          example: "177"
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/MessageAttachment'
  securitySchemes: { }